requirements/requirements_minigrid_memory.txt
# python 3.9
pip==21
setuptools==65.5.1
wheel==0.38.0
gymnasium==0.29.0
docopt==0.6.2
einops==0.4.1
Pillow==8.2.0 # ?
gym==0.18.3
gym-minigrid==1.0.2
moviepy==1.0.3
matplotlib==3.5
numpy==1.23.5
ruamel.yaml==0.18.10 # If you have problems installing via pip, try installing via conda

"""
# === Follow the order! ===
conda create -n rate_minigrid python=3.9
conda activate rate_minigrid
pip install -e .
pip install pip==21
pip install setuptools==65.5.1
pip install wheel==0.38.0
pip install gymnasium==0.29.0
pip install docopt==0.6.2
pip install einops==0.4.1
pip install Pillow==8.2.0
pip install gym==0.18.3
pip install gym-minigrid==1.0.2
pip install moviepy==1.0.3
pip install matplotlib==3.5
pip install numpy==1.23.5
pip install ruamel.yaml==0.18.10
"""